zeromq-scala-binding: Add a jzmq like Java API
Existing Java and Scala code that uses jzmq [1] can switch over to use
ZeroMQ over JNA by adopting the Scala binding for ZeroMQ.

Currently, the ZeroMQ Scala Binding does not support ZeroMQ 3.0 and
therefore the JZMQ like API cannot be used with ZeroMQ 3.0. Also, most
of the argument validation done in JZMQ is not implemented and
Poller#poll is not yet implemented.

1. https://github.com/zeromq/jzmq

Signed-off-by: Karim Osman <karim@iki.fi>